% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensicR-package.R
\docType{package}
\name{forensicR-package}
\alias{forensicR}
\alias{forensicR-package}
\title{forensicR: Crime Scene Documentation and Reconstruction Tools}
\description{
Supports crime scene investigators in documenting and reconstructing scenes. Scene points are mapped from field measurements (baseline, triangulation, polar). Shooting incidents are reconstructed from bullet defects with explicit uncertainty, following the methods described in Haag and Haag (2020, ISBN:9780128193976) and Hueske (2015, ISBN:9781498707664). Evidence and custody are logged in a structured form, and reproducible scene reports are rendered to Word, PDF or HTML.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kaquadros/forensicR}
  \item Report bugs at \url{https://github.com/kaquadros/forensicR/issues}
}

}
\author{
\strong{Maintainer}: Karine Quadros \email{kquadros@dgso.org}

Authors:
\itemize{
  \item Karine Quadros \email{kquadros@dgso.org}
  \item Allan Quadros \email{allanvcq@gmail.com}
}

}
\keyword{internal}
