% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furniture.R
\name{furniture_table}
\alias{furniture_table}
\title{Summary table of scene objects for reports}
\usage{
furniture_table(f)
}
\arguments{
\item{f}{A \code{furniture} tibble.}
}
\value{
A tibble with readable columns.
}
\description{
Summary table of scene objects for reports
}
