% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forensicR-package.R
\docType{package}
\name{forensicR-package}
\alias{forensicR}
\alias{forensicR-package}
\title{forensicR: Crime Scene Documentation and Reconstruction Tools}
\description{
Tools for crime scene investigators: scene mapping from field measurements (baseline, triangulation, polar), shooting reconstruction from bullet defects, structured evidence and custody logging, and reproducible scene reports rendered to Word, PDF or HTML.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kaquadros/forensicR}
  \item Report bugs at \url{https://github.com/kaquadros/forensicR/issues}
}

}
\author{
\strong{Maintainer}: Karine Quadros \email{kquadros@dgso.org}

Authors:
\itemize{
  \item Karine Quadros \email{kquadros@dgso.org}
  \item Allan Quadros \email{allanvcq@gmail.com}
}

}
\keyword{internal}
